Disable bootstrap classpath warning
Oracle javac 1.7 has a new warning:
    warning: [options] bootstrap class path not set in conjunction with
    -source 1.5

We're not going to hardcode a Java 1.5 bootstrap classpath to heed the
intent of this warning.  Continuing to declare source="1.5" seems
worthwhile for catching major source-compatibility mistakes.
target="1.5" doesn't seem to hurt, and anyway setting the property is
strongly recommended by the Ant documentation.  So, even though it's
insufficient for Java 1.5 binary compatibility, we're left with the
status quo.  Disable the warning.