workflows: automatically create GitHub release when tagged
Run `meson dist` on Linux in every build, because it's a good idea to test
this and because the build job already installs the dependencies for
`meson setup`.  Archive the artifact, then conditionally run a second job
to do the release.

This uses the same release-job pattern as OpenSlide Python, rather than
starting a second workflow upon completion of the build workflow.  This
results in a skipped job in every PR's job list, but allows the passing of
output variables between jobs.

Linux builds run in ubuntu-latest, not in a Fedora container, so the
Autotools blobs in the source tarball will be generated on Ubuntu 22.04.
That release has current versions of Autoconf and Automake, so this seems
okay.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>