Remove Autotools+Ant build system
There are some structural changes planned for the library, and having two
parallel build systems would make that harder.  We've carried both systems
in parallel for two releases / 15 months, so it should be reasonably
possible to drop the legacy one.

On platforms with older Meson and Python >= 3.7, it's still possible to
install OpenSlide Java by first upgrading Meson with `pip install
--upgrade meson`.

Closes: https://github.com/openslide/openslide-java/issues/45
Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>