workflows: remove source tarball as release artifact
The source tarball is now identical to a git-archive of the Git tag, so
we can defer to the automatic "Source code" links in the GitHub release
instead of uploading a separate artifact.

Drop "Run test build" from the release checklist, since CI will do that
for us on each PR, including the one in the next step.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>