Support OpenSlide cache management API
Use the same semantics as the OpenSlide API.

The OpenSlide class implements java.io.Closeable, which is documented to
be for "a source or destination of data that can be closed".
OpenSlideCache is technically a source or destination of data, but not at
the Java API level, and the only reason it needs to be closeable is to
release the underlying resource in the C API.  Its superinterface
java.lang.AutoCloseable is documented to be for "an object that may hold
resources [...] until it is closed"; implement that instead.

OpenSlide has a legacy dispose() method which is an alias for close().
Don't add a similar method to OpenSlideCache.

Closes: https://github.com/openslide/openslide-java/issues/36
Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>