Push locking down into OpenSlideFFM
Implement read-write locking for use-after-free checks in a single place
in OpenSlideFFM.Ref.  This is logically a better place for it and
simplifies the higher-level code.

Deprecate the now-unused, but public, zero-argument
OpenSlideDisposedException constructor.

OpenSlideFFM doesn't use the scoped locks inside the try-with-resources
blocks, and the compiler wants to warn about it.  Rather than add no-op
lock references or suppression boilerplate in every method, suppress try
warnings for the whole class.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>