Use AutoCloseable wrapper class for OpenSlide error checks
Rather than manually calling checkError() after every set of OpenSlide
calls, introduce a wrapper class which holds the OpenSlideRef and can be
used as a resource in a try-with-resources block.  When exiting the block,
its close() method can then run the error check.  This effectively uses
AutoCloseable as a Python context manager, which is a bit unusual but
doesn't violate the AutoCloseable API contract.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>